Centralizes event publication
The general architecture here is to use subdags to modularize the
different events that we publish.

A new version of `pipe-tools` is used as well. It includes a new feature
which allows us to define a base config when building subpipelines. This
let's us configure generic stuff on the parent pipeline (such as project
id, standard buckets and so on) while being able to override some of
those or add new config options in the sub pipelines.

The fishing events pipeline requires us to parametrize
the inference buffer days, a setting that's also used in
`pipe-features`. To do this, we'll pull directly from the global
`INFERENCE_BUFFER_DAYS` airflow variable. We can't use the general
strategy we do with other global variables where we use the `{{
var.value.GLOBAL_VARIABLE }}` value in a local
`pipe_events.fishing.local_variable` variable here, since we use this
parameter when building the start and end dates of the task and this is
already being interpolated. This means that we would need to recursively
interpolate variables, and airflow doesn't do that.