Fixes pipe events bigquery publication
* Adds a geography column to event tables

* Makes daily and monthly event pipelines use the same config variables

* Fixes time partitioning field in new bigquery schema

* Fixes delete daterange query which is using the old timestamp field

* The source table for encounter events is not date partitioned, so the
standard table sensor does not work. We are currently re-running the
entire thing each day for encounters, so a source table sensor is not
really needed.

* Removes date partitioninng from encounters events table. The encounter
events are re-generated every day with all the data, so date
partitioning fails because multiple partitions are updated every day. We
are disabling date paritioning in the tables for now.