Adds carrier portal queries
This is a squash commit which adds the final carrier portal queries to
this repo. It also ports a lot of the carrier portal specific features
to the general queries, such as port naming, regional information in
event info an so on.

Tailors the queries to generate the carrier portal dataset

Scopes the event queries to vessels involved in encountes with a carrier

Adds authorization status

Fixes broken region matching

Latest round of QA changes to the encounters query

Adds main and encountered vessel ports to encounters

Fixes anchorages query which was overwritten with the encounters query

Removes the event_geography column which we are not using and is super expensive to keep

Updates carrier portal queries to use latest vessel database

Fixes join logic being broken on encounters due to vessel database schema changes

Adds detailed authorization fields to encounter events

Adds TWN3 registry to the list of registries being processed

Generalizes anchorage and encounter changes