Changes the insertion logic from INSERT INTO to destination table
Using INSERT INTO is dangerous because the fields are matched using
numerical order, which means that if the output isn't in the same order
as the INSERT INTO fields it insert data in the wrong fields. Instead of
that, we are now using `bq query --destination_table`, which matches
fields using name instead.